+++ extend/bajo/hook/dobo@before-find-record.js
@@ -4,11 +4,22 @@ export async function rebuildFilter (modelName, filter, req) {
   const { isEmpty, isPlainObject, map, find, get } = this.app.lib._
   filter.query = filter.query ?? {}
 
-  const queryByModel = (query) => {
-    // by model
+  const queryBySiteSetting = (query) => {
+    if (!req.site) return
     const setting = get(req, `site.setting.dobo.query.${modelName}`)
     if (isPlainObject(setting) && !isEmpty(setting)) query.$and.push(setting)
-    return query
+  }
+
+  const queryByTeamSetting = (query) => {
+    if (!req.user) return
+    const q = []
+    for (const team of req.user.teams) {
+      const item = get(team, `setting.dobo.query.${modelName}`)
+      if (item) q.push(item)
+    }
+    if (isEmpty(q)) return
+    if (q.length === 1) query.$and.push(q[0])
+    else query.$and.push({ $or: q })
   }
 
   const model = this.app.dobo.getModel(modelName)
@@ -17,14 +28,17 @@ export async function rebuildFilter (modelName, filter, req) {
   const hasTeamId = model.hasProperty('teamId')
   const isAdmin = find(get(req, 'user.teams', []), { alias: 'administrator' }) && useAdmin.includes(get(req, 'routeOptions.config.ns'))
   const q = { $and: [] }
+  queryBySiteSetting(q)
+  queryByTeamSetting(q)
   if (!isEmpty(filter.query)) {
     if (filter.query.$and) q.$and.push(...filter.query.$and)
     else q.$and.push(filter.query)
   }
+  /*
   if (!(hasSiteId || hasUserId || hasTeamId)) {
-    filter.query = queryByModel(q)
     return filter
   }
+  */
   if (hasSiteId) q.$and.push({ siteId: req.site.id })
   if (hasTeamId && !isAdmin) {
     const teamIds = map(req.user.teams, 'id')
@@ -33,14 +47,13 @@ export async function rebuildFilter (modelName, filter, req) {
   } else if (!isAdmin) {
     if (hasUserId) q.$and.push({ userId: req.user.id })
   }
-  filter.query = queryByModel(q)
-  return filter
+  filter.query = q
 }
 
 export async function handler (modelName, filter, options = {}) {
   const { req } = options
   if (options.noAutoFilter || !req) return
-  filter = await rebuildFilter.call(this, modelName, filter, req)
+  await rebuildFilter.call(this, modelName, filter, req)
 }
 
 const doboBeforeFindRecord = {