+++ extend/bajo/hook/waibu@pre-parsing.js
@@ -1,8 +1,7 @@
 const preParsing = {
   level: 10,
   handler: async function (req, reply) {
-    const { getHostname } = this.app.waibu
-    req.site = await this.getSite(getHostname(req))
+    req.site = await this.getSite(req)
   }
 }